Update README.md
temperature converter web application created using HTML, CSS, and JavaScript. The application allows users to input a temperature and convert it between Celsius and Fahrenheit instantly. It's a simple and efficient tool for temperature conversions, making it easy for users to switch between different temperature scales with just a few clicks."